♻️ refactor: 파일 이동 및 타입 분리
